[mqtt] using infra log as default

+++ src/mqtt/mqtt_wrapper.h
@@ -10,7 +10,7 @@ int8_t HAL_GetDeviceSecret(char device_secret[IOTX_DEVICE_SECRET_LEN]);
 void *HAL_Malloc(uint32_t size);
 void HAL_Free(void *ptr);
 void HAL_Printf(const char *fmt, ...);
-int8_t HAL_Snprintf(char *str, const int len, const char *fmt, ...);
+int HAL_Snprintf(char *str, const int len, const char *fmt, ...);
 uint64_t HAL_UptimeMs(void);
 void HAL_Srandom(uint32_t seed);
 void *HAL_MutexCreate(void);
